Fix showing updated profile image
When updating a profile image of the user, the old profile image was still shown and the user must reload the page so that the changes could be visible.

To fix this issue, first introduce an appContext component that exposes an userState that can be set by any of its children. Let the profile component update the userState as soon as a profile image gets updated.